Replace FFmpeg.AutoGen with ClangSharp generated binding.

+++ osu.Framework/Graphics/OpenGL/Textures/GLVideoTexture.cs
@@ -72,7 +72,7 @@ namespace osu.Framework.Graphics.OpenGL.Textures
                 GL.PixelStore(PixelStoreParameter.UnpackRowLength, videoUpload.Frame->linesize[i]);
 
                 GL.TexSubImage2D(TextureTarget2d.Texture2D, 0, 0, 0, videoUpload.GetPlaneWidth(i), videoUpload.GetPlaneHeight(i),
-                    PixelFormat.Red, PixelType.UnsignedByte, (IntPtr)videoUpload.Frame->data[i]);
+                    PixelFormat.Red, PixelType.UnsignedByte, (IntPtr)videoUpload.Frame->data[(int)i]);
 
                 GL.PixelStore(PixelStoreParameter.UnpackRowLength, 0);
             }
